Allows the use of the API_PATH environment variable in CI (#179)
In some CI environments, services running inside a container cannot be
accessed via the ip 127.0.0.1. Another host must be used. So we need to
"inform" Volto and Cypress of the new Plone host that is running in a
container. Volto's Cypress tests allow this with the API_PATH
environment variable. But Cypress tests do not inherit the OS
environment variables. We need to pass the API_PATH variable explicitly.

If there is no API_PATH environment variable,
http://127.0.0.1:55001/plone is used.