Fix Eslint in IDEs
The generated package includes a `babel.config.js` file that requires
the `babel-plugin-react-intl` and `babel-preset-razzle` dependencies.
These dependencies are not declared in the package, so they are neither
in the root node_modules folder nor in the package's own node_modules
folder. When some IDEs run ESLint, they don't recognize that these
packages are located in the .pnpm subfolder of the root node_modules,
which causes a "package not found" error.

When the packages are declared as dependencies, they are placed in the
package's node_modules folder and are correctly resolved by IDEs.