Fix Eslint in IDEs
The generated package includes a babel.config.js file that requires the
babel-plugin-react-intl and babel-preset-razzle dependencies. These
dependencies are not declared in the package, so they are neither in the
root node_modules folder nor in the package's own node_modules folder.
When some IDEs run ESLint, they don't recognize that these packages are
located in the .pnpm subfolder of the root node_modules, which causes a
"package not found" error.

So we hoist these packages. This causes them to be placed at the root of
the `node_modules` folder in the monorepo root. This way, IDEs’ ESLint
can see the packages, and the error no longer occurs.