Respects Python version when installing backend in project
The backend addon's `@uv venv $(VENV_FOLDER)` command respects the Python version in pyproject.toml.
But the project's `make install` command is executed in the context of the project root, where we don't have pyproject.toml.
This did that the version of Python used was the system version, not the version required by the project.
So we need to enter the backend folder, so that `uv` fills pyproject.toml.